w_wad: widen the wad file size, and refuse what the format cannot describe
filestream_get_size returns int64_t and W_AddFile stored it in an int.
The narrowing was silent, and the result went on to mmap(), malloc() and
munmap() as a size_t.  Measured on a 2GiB file: the field held
-2147483632, which as a size_t is 18446744071562067984 - so both the
mapping and the allocation were being asked for eighteen exabytes, and
the interesting behaviour was in whichever failure path ran first.

wadfile_info_t::length is int64_t now, and so are the things the file
size actually flows into: W_IsPK3's length, W_TranslatePK3's zip_length
and out_length, and newlen.

Deliberately not widened.  The local `length` in W_AddFile is the lump
directory's size, numlumps * sizeof(filelump_t), bounded by the format's
32-bit numlumps - it is not the file size and widening it would only
blur what the type means.  filelump_t::size stays int because it is the
on-disk record, so the single-lump path casts explicitly and is safe
because of the bound below.

The bound is the real fix.  A wad's directory offsets are 32-bit by
specification, so a file past 2GB is malformed rather than merely large;
W_AddFile now refuses it by name instead of narrowing it.  W_TranslatePK3
refuses the same way for its own reason - the zip structures it reads
are 32-bit and ZIP64 is not implemented here, so an oversized archive is
rejected rather than passed down and narrowed into pk3_zip_open's int.

Verified against real content rather than by reasoning.  Eleven files -
DOOM, Doom2, Heretic, Hexen, Chex 1 and 3, Avulsion, Biodiverse,
Dwelling at 101MB, and the myhouse, zdcmp2 and hdoom PK3s - identify and
translate byte-identically before and after the change: myhouse still
inflates 68881167 to 144304079, zdcmp2 to 102124514, hdoom to 82613274.
The widening changes nothing that works today, which is the whole claim.

Builds clean with no warnings, which matters here because a width
mismatch in a comparison or a format string is exactly what would show
up as one.

+++ src/w_pk3.h
@@ -9,18 +9,20 @@
  * served from a memory buffer). */
 
 #ifndef __W_PK3__
+#include <stdint.h>
+
 #define __W_PK3__
 
 #include "doomtype.h"
 
 /* ZIP local-file-header magic check */
-dbool W_IsPK3(const unsigned char *data, int length);
+dbool W_IsPK3(const unsigned char *data, int64_t length);
 
 /* Translate a ZIP archive in memory into a synthesized PWAD image.
  * Returns a malloc'd buffer (caller owns) and its size in *out_length,
  * or NULL on a malformed archive. */
-unsigned char *W_TranslatePK3(const unsigned char *zip, int zip_length,
-                              int *out_length, const char *archive_name);
+unsigned char *W_TranslatePK3(const unsigned char *zip, int64_t zip_length,
+                              int64_t *out_length, const char *archive_name);
 
 /* Global lump number for an exact archive path (e.g. "decorate/monster/
  * imp.txt"), disambiguating same-basename lumps that collide in the 8-char
